backend/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3
python-decouple==3.8
coreapi  # для drf-spectacular (альтернатива)
drf-spectacular  # современная замена drf-yasg



